+++ tree/phylotree.h
@@ -108,14 +108,14 @@ inline T *aligned_alloc(size_t size) {
 #if (defined(__GNUC__) || defined(__clang__)) && !defined(WIN32) &&!defined(WIN64) && !defined(__CYGWIN__)
         print_stacktrace(cerr);
 #endif
-        outError("Not enough memory, allocation of " + convertInt64ToString(size*sizeof(T)) + " bytes failed (bad_alloc)");
+        outError("Not enough memory, allocation of " + convertInt64ToString(static_cast<int64_t>(size*sizeof(T))) + " bytes failed (bad_alloc)");
     }
 #endif
     if (mem == nullptr) {
 #if (defined(__GNUC__) || defined(__clang__)) && !defined(WIN32) &&!defined(WIN64) && !defined(__CYGWIN__)
         print_stacktrace(cerr);
 #endif
-        outError("Not enough memory, allocation of " + convertInt64ToString(size*sizeof(T)) + " bytes failed (bad_alloc)");
+        outError("Not enough memory, allocation of " + convertInt64ToString(static_cast<int64_t>(size*sizeof(T))) + " bytes failed (bad_alloc)");
     }
     return (T*)mem;
 }
